Moved and renamed management view models
