--- GIFrameworkMap.Data/Data/Models/ViewModels/Management/LayerSourceOptionEditModel.cs
@@ -1,8 +1,0 @@
-﻿namespace GIFrameworkMaps.Data.Models.ViewModels.Management
-{
-	public class LayerSourceOptionEditModel
-    {
-        public LayerSourceOption? LayerSourceOption { get; set; }
-        public LayerSource? LayerSource { get; set; }
-    }
-}
