+++ GIFrameworkMap.Data/ViewModels/Management/AnalyticsViewModel.cs
@@ -1,0 +1,11 @@
+﻿using GIFrameworkMaps.Data.Models;
+using System.Collections.Generic;
+
+namespace GIFrameworkMaps.Data.ViewModels.Management
+{
+	public class AnalyticsViewModel
+	{
+		public IList<AnalyticsDefinition> AvailableAnalytics { get; set; } = [];
+		public int VersionID { get; set; }
+	}
+}
