+++ GIFrameworkMap.Data/ViewModels/Management/LayerSourceOptionEditViewModel.cs
@@ -1,0 +1,10 @@
+﻿using GIFrameworkMaps.Data.Models;
+
+namespace GIFrameworkMaps.Data.ViewModels.Management
+{
+	public class LayerSourceOptionEditViewModel
+	{
+		public LayerSourceOption? LayerSourceOption { get; set; }
+		public LayerSource? LayerSource { get; set; }
+	}
+}
